[ARV-57] test : test support 통함
