--- src/test/java/com/backend/allreva/ControllerTestSupport.java
@@ -1,14 +1,0 @@
-package com.backend.allreva;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.test.web.servlet.MockMvc;
-
-@AutoConfigureMockMvc
-public abstract class ControllerTestSupport {
-    @Autowired
-    protected MockMvc mockMvc;
-    @Autowired
-    protected ObjectMapper objectMapper;
-}
